requirements.txt
# Django dependencies

django
django-ninja
django-ninja-extra
django-ninja-jwt

# AI Agent dependencies

ollama
pydantic-ai

# DB dependencies

psycopg2-binary
pgvector

# Android notification dependencies

firebase-admin

# Gunicorn for Docker deployment

gunicorn

# Taks queue dependencies

django-q2